requirements.txt
PySide6==6.6.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.13
groq==0.32.0

# Google TTS (optional, for better voice quality)
gTTS==2.5.0
pygame==2.5.2

# Face tracking with MediaPipe
mediapipe==0.10.14
opencv-contrib-python==4.12.0.88

# Servo motor control libraries
RPi.GPIO==0.7.1
gpiozero==2.0.1
pigpio==1.78
adafruit-circuitpython-servokit==1.3.21
adafruit-pca9685==1.0.1
